+++ libs/qsynedit/qsynedit/formatter/cppformatter.cpp
@@ -42,21 +42,19 @@ namespace QSynedit {
             if (rangePreceeding.state == CppSyntaxer::RangeState::rsRawStringNotEscaping)
                 return 0;
 
-            // } can't find the correct indent , if autoindent for ( is handled here.
-            // We need to find a way to find the correct start line of { (
-            // if (rangePreceeding.getLastIndentType() == IndentType::Parenthesis) {
-            //     bool lastLineHasLastParentheis = true;
-            //     if (startLine > 1) {
-            //         SyntaxState synState = editor->document()->getSyntaxState(startLine - 2);
-            //         lastLineHasLastParentheis = rangePreceeding.parenthesisLevel > synState.parenthesisLevel;
-            //     }
-            //     if (lastLineHasLastParentheis) {
-            //         indentSpaces = findLastParenthesis(startLine, editor);
-            //         if (addIndent)
-            //             indentSpaces += 1;
-            //     }
-            //     return indentSpaces;
-            // }
+            if (rangePreceeding.getLastIndentType() == IndentType::Parenthesis) {
+                bool lastLineHasLastParentheis = true;
+                if (startLine > 1) {
+                    SyntaxState synState = editor->document()->getSyntaxState(startLine - 2);
+                    lastLineHasLastParentheis = rangePreceeding.parenthesisLevel > synState.parenthesisLevel;
+                }
+                if (lastLineHasLastParentheis) {
+                    indentSpaces = findLastParenthesis(startLine, editor);
+                    if (addIndent)
+                        indentSpaces += 1;
+                }
+                return indentSpaces;
+            }
             if (addIndent) {
     //            QString trimmedS = s.trimmed();
                 QString trimmedLineText = lineText.trimmed();
@@ -158,22 +156,24 @@ namespace QSynedit {
     {
     }
 
-    // int CppFormatter::findLastParenthesis(int line, const QSynEdit *editor)
-    // {
-    //     if (line == 1)
-    //         editor->syntaxer()->resetState();
-    //     else {
-    //         SyntaxState rangeState = editor->document()->getSyntaxState(line-2);
-    //         editor->syntaxer()->setState(rangeState);
-    //     }
-    //     editor->syntaxer()->setLine(editor->lineText(line),line);
-    //     int pos = 0;
-    //     while (!editor->syntaxer()->eol()) {
-    //         if (editor->syntaxer()->getTokenAttribute() == editor->syntaxer()->symbolAttribute()
-    //             && editor->syntaxer()->getToken() == "(")
-    //             pos = editor->syntaxer()->getTokenPos();
-    //         editor->syntaxer()->next();
-    //     }
-    //     return pos;
-    // }
+    int CppFormatter::findLastParenthesis(int line, const QSynEdit *editor)
+    {
+        if (line == 1)
+            editor->syntaxer()->resetState();
+        else {
+            SyntaxState rangeState = editor->document()->getSyntaxState(line-2);
+            editor->syntaxer()->setState(rangeState);
+        }
+        QString lineText = editor->lineText(line);
+        int leading = editor->leftSpaces(lineText);
+        editor->syntaxer()->setLine(lineText.trimmed(),line);
+        int pos = 0;
+        while (!editor->syntaxer()->eol()) {
+            if (editor->syntaxer()->getTokenAttribute() == editor->syntaxer()->symbolAttribute()
+                && editor->syntaxer()->getToken() == "(")
+                pos = editor->syntaxer()->getTokenPos();
+            editor->syntaxer()->next();
+        }
+        return leading+pos;
+    }
 }