+++ libs/qsynedit/qsynedit/formatter/cppformatter.h
@@ -20,7 +20,7 @@ public:
 protected:
     int findCommentStartLine(int searchStartLine, const QSynEdit *editor);
     void doInitOptions() override;
-    // int findLastParenthesis(int line, const QSynEdit *editor);
+    int findLastParenthesis(int line, const QSynEdit *editor);
 };
 }