+++ NEWS.md
@@ -32,6 +32,7 @@ Red Panda C++ Version 3.2
   - enhancement: Display a warning message if no compiler set is found at the first time RedPanda C++ runs.
   - fix: Can't find the single word if the caret is inside it.
   - fix: Search Dialog always returns to center each time starting a search.
+  - enhancement: Auto indent contents in parenthesis.
 
 Red Panda C++ Version 3.1